requirements.txt
django
djangorestframework
environs
pgvector
django-guardian